LayoutTests/imported/w3c/web-platform-tests/html/syntax/speculative-parsing/generated/page-load/math-script-src.tentative-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_equals: speculative case incorrectly fetched expected "" but got "param-encodingcheck: %C4%9E\r\nAccept: */*\r\nReferer: http://localhost:8800/html/syntax/speculative-parsing/generated/p...


FAIL Speculative parsing, page load: math-script-src Unhandled rejection: assert_equals: speculative case incorrectly fetched expected "" but got "param-encodingcheck: %C4%9E\r\nAccept: */*\r\nReferer: http://localhost:8800/html/syntax/speculative-parsing/generated/page-load/resources/math-script-src-framed.sub.html?uuid=77933028-62ef-4d7d-b1a1-c61864746ae1"
